Add user registration and enhance login functionality

+++ base/views.py
@@ -5,6 +5,7 @@ from django.contrib.auth.models import User
 from .models import Room,Topic
 from .forms import RoomForm
 from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
 from django.contrib.auth import authenticate,login,logout
 # Create your views here.
@@ -16,8 +17,12 @@ from django.contrib.auth import authenticate,login,logout
 # ]
 
 def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method=='POST':
-        username=request.POST.get('username')
+        username=request.POST.get('username').lower()
         password=request.POST.get('password')
         try:
             user=User.objects.get(username=username)
@@ -32,12 +37,26 @@ def loginPage(request):
         else:
             messages.error(request,'Username OR password is incorrect')
         
-    context={}
+    context={'page':page}
     return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form=UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error has occured during registration')
+    return render(request,'base/login_register.html',{'form':form})
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q') != None else ''
@@ -76,7 +95,7 @@ def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
 
-    if request.user != room.user:
+    if request.user != room.host :
         return HttpResponse('You are not allowed here!')
 
     if request.method=='POST':
@@ -91,6 +110,9 @@ def updateRoom(request,pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host :
+        return HttpResponse('You are not allowed here!')
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
